edit read bin file

+++ cf-vae/train_img_vae.py
@@ -7,7 +7,7 @@ from scipy.sparse import load_npz
 np.random.seed(0)
 tf.set_random_seed(0)
 
-images = np.fromfile("data/amazon/images.bin")
+images = np.fromfile("data/amazon/images.bin", dtype=np.uint8)
 data = images.reshape((16000, 3072))
 
 idx = np.random.rand(data.shape[0]) < 0.8
